norme

+++ srcs/routine_exit.c
@@ -3,23 +3,27 @@
 #include "exec.h"
 #include <signal.h>
 
-void	kill_job_active(t_list *lst)
+void	kill_job_stopped(t_list *lst)
 {
 	t_job	*j;
 
 	while (lst)
 	{
 		j = lst->data;
-		kill(-j->pgid, SIGKILL);
+		if (j->status == STOPPED)
+			kill(-j->pgid, SIGKILL);
 		lst = lst->next;
 	}
 }
 
 void	exit_routine(t_cfg *shell, uint8_t ret)
 {
-	kill_job_active(shell->job);
-	if (shell->interactive)
+	kill_job_stopped(shell->job);
+	if (shell->interactive &&
+			(int)shell->term_origin.c_iflag != -1)
 		set_termios(TCSADRAIN, &shell->term_origin);
+	close(shell->debug);
+	close(FD_NONINT);
 	clean_cfg(shell);
 	exit(ret);
 }
